fix bags
