+++ fibonacci/answers/brute_force_fibonacci.py
@@ -1,4 +1,4 @@
-class Fibonacci(object):
+class BruteForceFibonacci(object):
 
     def __init__(self):
         self.cache = {}
@@ -23,20 +23,3 @@ class Fibonacci(object):
             self.cache[n] = ans
 
         return ans
-
-    def dynamic_implementation(self, n):
-        """
-        :param n: n-th number of fibonacci numbers
-        :return: fibonacci number
-        :rtype: int
-        """
-        if n < 0:
-            raise TypeError('please enter positive numbers')
-
-        self.cache[0] = 0
-        if n > 0:
-            self.cache[1] = 1
-            for i in range(2, n + 1):
-                self.cache[i] = self.cache[i - 1] + self.cache[i - 2]
-
-        return self.cache[n]
