+++ fibonacci/dynamic_fibonacci.py
@@ -5,11 +5,9 @@ The simplest is the series 1, 1, 2, 3, 5, 8, etc.
 """
 
 
-class Fibonacci(object):
+class DynamicFibonacci(object):
     """
-    In this Fibonacci class, we realized two implementations of Fibonacci numbers.
-    First one is recursive implementation.
-    The second one is dynamic implementation.
+    In this DynamicFibonacci class, we realized dynamic implementation.
     """
 
     def __init__(self):
@@ -18,20 +16,6 @@ class Fibonacci(object):
         instance.
 
         Here you need initialize some dictionary variable which will be store values of fibonacci.
-        """
-        pass
-    def recursive_implementation(self, n):
-        """
-        :param n: n-th number of fibonacci numbers
-        :rtype: int
-        :return: fibonacci number
-
-        Recursion means calling functions directly from itself.
-
-        You must determine the number n is negative or positive. If the number is negative, then you need to raise the error.
-        After that you need to check if this number is in the cache. If this number in the cache return this number.
-        If not, you need to recursively find (n - 1) + (n - 2).
-        Until n equals one or zero.
         """
         pass
